refactor(settings): extract Screen into its own top-level section
The Screen controls (UI scale, orientation, fullscreen, keep screen
on, dock position, assistant launch) lived as a SettingsSubheader
group inside the still-inline Display section. Move the six rows and
their private Boolean.toFullscreen() helper verbatim into a new
components/ScreenSection.kt, and call it as its own top-level section
between Display and Units.

Rename the settings_subheader_screen string resource to
settings_section_screen (same "Screen" copy) since it now titles a
full section instead of a subheader nested inside Display.

Pure structural move: no behavior change. Display still holds
Theme/Accent/Preset/Glass/Fonts; a later task finishes splitting it
into Appearance.
